// lib/api/generated/schemas/adminShopBannersControllerListStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * НеМалика API
 * Маркетплейс компьютерной техники — REST API для веб-витрины (Next.js) и Telegram mini-app. Покупать можно без авторизации; вход через Telegram нужен продавцу, администратору и покупателю, который хочет видеть историю просмотров на всех своих устройствах.
 * OpenAPI spec version: 1.0.0
 */

export type AdminShopBannersControllerListStatus = typeof AdminShopBannersControllerListStatus[keyof typeof AdminShopBannersControllerListStatus];


export const AdminShopBannersControllerListStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;